add(view_project): view project page and provide its bloc
